🐛채팅방 리스트 SSE 수정

+++ app/src/main/java/com/project/heyboardgame/dataModel/ChatListResult.kt
@@ -9,7 +9,7 @@ data class ChatListResult (
 
 data class ChatRoom (
     @SerializedName("id")
-    val roomId: Int,
+    val roomId: Long,
     @SerializedName("userInfo")
     val userInfo: Friend,
     @SerializedName("lastMessage")
